+++ govtool/frontend/src/consts/dRepActions/fields.ts
@@ -1,5 +1,10 @@
 import i18n from "@/i18n";
-import { URL_REGEX, isReceivingAddress, isValidURLLength } from "@/utils";
+import {
+  IMAGE_REGEX,
+  URL_REGEX,
+  isReceivingAddress,
+  isValidURLLength,
+} from "@/utils";
 
 export const Rules = {
   GIVEN_NAME: {
@@ -64,6 +69,12 @@ export const Rules = {
       message: i18n.t("registration.fields.validations.maxLength", {
         maxLength: 1000,
       }),
+    },
+  },
+  IMAGE_URL: {
+    pattern: {
+      value: IMAGE_REGEX,
+      message: i18n.t("registration.fields.validations.image"),
     },
   },
 };