+++ CHANGELOG.md
@@ -12,6 +12,8 @@ changes.
 
 ### Added
 
+- Add image tag for DRep in GovTool metadata [3137](https://github.com/IntersectMBO/govtool/issues/3137)
+
 ### Fixed
 
 ### Changed
@@ -19,7 +21,6 @@ changes.
 ### Removed
 
 ## [v2.0.14](https://github.com/IntersectMBO/govtool/releases/tag/v2.0.14) 2025-03-05
-
 
 ### Added